Delete should not replay tag writes
Regs #335
